Change the 2LS reflected fault op to take a ctx
VM exits are another type of reflected fault.  The existing reflected fault
code assumes they are hardware faults.  For the most part, the handling of
the two types of contexts is the same, as far as uthread.c is concerned.

By passing the context, instead of interpretting the context, directly to
the 2LS, we'll eventually be able to handle VM exits as well, without the
parlib / uthread code caring at all.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>