A virtio-mmio milestone
This is a big milestone for the series of commits (starting with
e1f911d) focused on cleaning up and improving our virtio-mmio implementation.

We are now in a much better place with virtio, and this should put Gan
on solid footing as he takes over and starts building our virtio-net
implementation.

There are still several TODOs in the code, but they are well commented.

The two largest remaining items are:
- Having the device set the DEVICE_NEEDS_RESET status bit in the event
  that it encounters an unrecoverable error; at present the device will
  crash the vm in the event of such an error. Note that the Linux driver
  does not, at present, even use this status bit (the bit macro was,
  in fact, only added to Linux about a month ago). It will probably
  also be helpful to have errors that just print a message and kill the
  vm for the time being. Fail-fast is good.
- Properly terminating and cleaning up the queue service threads when
  either a device reset is issued or the driver disables a queue. There
  are comments in the code to this effect.

Signed-off-by: Michael Taufen <mtaufen@gmail.com>